Move externals definitions to a new file
