+++ src/sagas/index.ts
@@ -9,8 +9,9 @@ import * as actions from '../actions'
 import * as actionTypes from '../actions/actionTypes'
 import { WorkspaceLocation } from '../actions/workspaces'
 import { mockBackendSaga } from '../mocks/backend'
-import { defaultEditorValue, externalLibraries, IState } from '../reducers/states'
 import { IVLE_KEY, USE_BACKEND } from '../utils/constants'
+import { externalLibraries } from '../reducers/externalLibraries'
+import { defaultEditorValue, IState } from '../reducers/states'
 import { showSuccessMessage, showWarningMessage } from '../utils/notification'
 import backendSaga from './backend'
